pyproject.toml
[project]
name = "luganda-mt-synthetic"
version = "0.1.0"
description = "Research pipeline for evaluating LLMs on Luganda->English machine translation."
readme = "README.md"
requires-python = "==3.11.*"
dependencies = [
    "torch==2.7.1",
    "transformers>=4.43.0",
    "bitsandbytes>=0.46.0",
    "accelerate>=0.31.0",
    "peft>=0.11.0",
    "unbabel-comet>=2.2.1",
    "mlflow>=2.13.0",
    "datasets>=2.20.0",
    "pandas>=2.2.0",
    "tqdm>=4.66.0",
    "sentencepiece>=0.2.0",
    "protobuf>=4.25.0",
    "huggingface-hub>=0.24.0",
    "python-dotenv>=1.0.0",
    "setuptools>=70.0.0,<81",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]

[tool.uv.sources]
torch = { index = "pytorch-cu128" }

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
